refactor(commons): :hammer: duplicate lines was refactored
