+++ packages/common/decorators/service.decorator.ts
@@ -1,13 +1,6 @@
-import "reflect-metadata";
-import { IDependenciesClass } from "../interfaces";
-import { Logger } from "../utils/logger.utils";
-import { unCapitalize, getConstructorParamNames } from "../utils/shared.utils";
-import {
-  DEPENDENCIES_CONSTRUCTOR,
-  DEPENDENCIES_CLASS,
-  IS_SERVICE,
-} from "../utils/constants";
-const logger = Logger();
+import { IS_SERVICE } from "../utils/constants";
+import { createClassDecorator } from "./create.decorator";
+
 /**
  * Decorador de clase para marcar una clase como inyectable.
  *
@@ -17,34 +10,5 @@ const logger = Logger();
  * @decorator
  * @returns {ClassDecorator} Una función de decorador de clase.
  */
-export const Injectable = (): ClassDecorator => {
-  return (target: Function) => {
-    logger.debug("@Service target", target.name);
-
-    /** Obtiene los tipos de dependencias de la clase utilizando los metadatos de diseño.*/
-    const dependencies: any[] =
-      Reflect.getMetadata(DEPENDENCIES_CONSTRUCTOR, target) ?? [];
-    logger.debug("@Service dependencies", dependencies);
-
-    /** Obtiene los nombres de los parámetros del constructor de la clase.*/
-    const paramNames: any[] = getConstructorParamNames(target);
-    logger.debug("@Service paramNames", paramNames);
-
-    /** Representa las dependencias de la clase en un formato específico.*/
-    const dependenciesClass: Array<IDependenciesClass> = dependencies.map(
-      (dependencie, index) => {
-        return {
-          type: "Service",
-          TargeName: target.name,
-          nameClass: dependencie.name,
-          nameClassContainer: unCapitalize(dependencie.name),
-          nameParameter: paramNames[index],
-        };
-      },
-    );
-
-    /** Define los metadatos de dependencias en la clase. */
-    Reflect.defineMetadata(DEPENDENCIES_CLASS, dependenciesClass, target);
-    Reflect.defineMetadata(IS_SERVICE, true, target);
-  };
-};
+export const Injectable = (): ClassDecorator =>
+  createClassDecorator("Service", IS_SERVICE);
