+++ packages/common/services/logger.user.service.ts
@@ -1,9 +1,9 @@
 /* eslint-disable no-console */
-import * as util from "util";
-import { ILoggerUserService, IOptionsLog } from "../interfaces";
-import { coerceBooleanProperty, colorPrint } from "../utils/shared.utils";
+import { ABSBaseLoggerService } from "./base.logger.service";
 import { ProcessDataService } from "./masker/process/processData.service";
 import { ProviderPatternService } from "./masker/process/providerPattern.service";
+import { ILoggerUserService, IOptionsLog } from "../interfaces";
+import { coerceBooleanProperty, colorPrint } from "../utils/shared.utils";
 
 /**
  * Servicio para manejar el registro de mensajes con diferentes niveles de importancia.
@@ -16,16 +16,14 @@ import { ProviderPatternService } from "./masker/process/providerPattern.service
  * logger.error('Mensaje de error');
  * ```
  */
-export class LoggerUserService implements ILoggerUserService {
+export class LoggerUserService
+  extends ABSBaseLoggerService
+  implements ILoggerUserService
+{
   /**
-   * Instancia única del servicio LoggerService.
+   * Instancia única del servicio LoggerUserService.
    */
   private static instance: LoggerUserService;
-  private readonly options: IOptionsLog = {
-    withColor: true,
-    activeMasker: false,
-    configSchemaMasker: {},
-  };
   private readonly providerPattern: ProviderPatternService =
     ProviderPatternService.getInstance();
   private readonly processData: ProcessDataService =
@@ -35,16 +33,14 @@ export class LoggerUserService implements ILoggerUserService {
    * Constructor privado para asegurar que no se pueda instanciar directamente.
    */
   private constructor(options?: IOptionsLog) {
-    this.options.withColor = options?.withColor ?? true;
-    this.options.activeMasker = options?.activeMasker ?? false;
-    this.options.configSchemaMasker = options?.configSchemaMasker ?? {};
+    super(options);
     this.initializeMasker();
   }
 
   /**
-   * Obtiene la única instancia de LoggerService.
+   * Obtiene la única instancia de LoggerUserService.
    * @param options - Listado de opciones para aplicar a la instancia.
-   * @returns La única instancia de LoggerService.
+   * @returns La única instancia de LoggerUserService.
    */
   static getInstance(options?: IOptionsLog): ILoggerUserService {
     if (!this.instance) {
@@ -53,7 +49,7 @@ export class LoggerUserService implements ILoggerUserService {
     return this.instance;
   }
 
-  private initializeMasker() {
+  protected initializeMasker() {
     if (this.options.activeMasker && this.options.configSchemaMasker) {
       this.providerPattern.setupSchema(this.options.configSchemaMasker);
       const keys = Object.keys(this.options.configSchemaMasker);
@@ -64,7 +60,7 @@ export class LoggerUserService implements ILoggerUserService {
     }
   }
 
-  private deselectSchema() {
+  protected deselectSchema() {
     if (this.enableDeselectSchema) {
       this.processData.deselectSchema();
     }
@@ -76,31 +72,6 @@ export class LoggerUserService implements ILoggerUserService {
   }
 
   /**
-   * Formatea el mensaje a ser registrado.
-   *
-   * @param level - Nivel del mensaje (info, error, etc.).
-   * @param message - Mensaje a ser formateado.
-   * @returns Mensaje formateado.
-   */
-  private formatMessage(level: string, message: any) {
-    return `[${level.toUpperCase()}]: ${message}`;
-  }
-
-  /**
-   * Formatea el argumento que se va a imprimir.
-   *
-   * @param arg - argumento a formatear para imprimir.
-   * @returns argumentos formateado.
-   */
-  private formatArg(arg: any) {
-    return util.inspect(arg, {
-      showHidden: false,
-      depth: null,
-      colors: this.options.withColor,
-    });
-  }
-
-  /**
    * Registra el mensaje en la consola.
    *
    * @param color - Color asociado al nivel de mensaje.
@@ -108,13 +79,13 @@ export class LoggerUserService implements ILoggerUserService {
    * @param message - Mensaje a ser registrado.
    * @param arg - Argumentos adicionales.
    */
-  private log(
+  protected log(
     color: string,
     level: string,
     message: any,
     arg: any,
     ...otherArg: any
-  ) {
+  ): void {
     // Si el logger no está activo, salir temprano
     if (!coerceBooleanProperty(process.env.ZANOBIJS_LOGGER_USER)) {
       return;
@@ -152,69 +123,4 @@ export class LoggerUserService implements ILoggerUserService {
     this.deselectSchema();
   }
 
-  /**
-   * Registra un mensaje informativo.
-   *
-   * @param message - Mensaje informativo.
-   * @param arg - Argumentos principal a imprimir.
-   * @param otherArgs - Argumentos adicionales.
-   */
-  info(message: string, arg: any, ...otherArgs: any) {
-    this.log(colorPrint.blue, "info", message, arg, ...otherArgs);
-  }
-
-  /**
-   * Registra una advertencia.
-   *
-   * @param message - Mensaje de advertencia.
-   * @param arg - Argumentos principal a imprimir.
-   * @param otherArgs - Argumentos adicionales.
-   */
-  warn(message: string, arg: any, ...otherArgs: any) {
-    this.log(colorPrint.orange, "warn", message, arg, ...otherArgs);
-  }
-
-  /**
-   * Registra un mensaje de error.
-   *
-   * @param message - Mensaje de error.
-   * @param arg - Argumentos principal a imprimir.
-   * @param otherArgs - Argumentos adicionales.
-   */
-  error(message: string, arg: any, ...otherArgs: any) {
-    this.log(colorPrint.red, "error", message, arg, ...otherArgs);
-  }
-
-  /**
-   * Registra un mensaje de éxito.
-   *
-   * @param message - Mensaje de éxito.
-   * @param arg - Argumentos principal a imprimir.
-   * @param otherArgs - Argumentos adicionales.
-   */
-  success(message: string, arg: any, ...otherArgs: any) {
-    this.log(colorPrint.green, "success", message, arg, ...otherArgs);
-  }
-
-  /**
-   * Registra un mensaje de depuración.
-   *
-   * @param message - Mensaje de depuración.
-   * @param arg - Argumentos principal a imprimir.
-   * @param otherArgs - Argumentos adicionales.
-   */
-  debug(message: string, arg: any, ...otherArgs: any) {
-    this.log(colorPrint.white, "debug", message, arg, ...otherArgs);
-  }
-
-  /**
-   * Registra un mensaje de depuración.
-   *
-   * @param message - Mensaje de depuración.
-   * @param arg - Argumentos principal a imprimir.
-   * @param otherArgs - Argumentos adicionales.
-   */
-  important(message: string, arg: any, ...otherArgs: any) {
-    this.log(colorPrint.BgRed, "important", message, arg, ...otherArgs);
-  }
 }
