+++ packages/common/services/masker/patterns/emailPattern.service.ts
@@ -13,10 +13,8 @@ class EmailPattern implements IPattern {
   }
 
   public mask(text: string): string {
-    // const emailRegex = /([a-zA-Z0-9._%+-]+)@([a-zA-Z0-9.-]+\.[a-zA-Z]{2,})/g;
     const emailRegex =
       /([a-zA-Z0-9._%+-]{1,64})@([a-zA-Z0-9.-]{1,255}\.[a-zA-Z]{2,})/g;
-    // const emailRegex = /^([a-zA-Z0-9._%+-]+)@([a-zA-Z0-9.-]+\.[a-zA-Z]{2,})$/g;
 
     return text.replace(emailRegex, (match, local, domain) => {
       const domainParts = domain.split(".");
